third_party/WebKit/LayoutTests/css3/blending/blend-mode-property-parsing-expected.txt

Test the parsing of the mix-blend-mode property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Basic reference : normal
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'normal'


Basic reference : multiply
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'multiply'


Basic reference : screen
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'screen'


Basic reference : overlay
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'overlay'


Basic reference : darken
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'darken'


Basic reference : lighten
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'lighten'


Basic reference : color-dodge
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color-dodge'


Basic reference : color-burn
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color-burn'


Basic reference : hard-light
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'hard-light'


Basic reference : soft-light
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'soft-light'


Basic reference : difference
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'difference'


Basic reference : exclusion
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'exclusion'


Basic reference : hue
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'hue'


Basic reference : saturation
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'saturation'


Basic reference : color
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'color'


Basic reference : luminosity
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('mix-blend-mode') is 'luminosity'
PASS successfullyParsed is true

TEST COMPLETE
